박훈나 리뷰, 영화삭제 수정 # WARNING: head commit changed in the meantime
Merge branch 'master' of
https://github.com/wooseongryu/class3_cinepick.git

+++ CinepickProject/src/main/java/com/itwillbs/cinepick/service/MovieService.java
@@ -42,6 +42,12 @@ public class MovieService {
 	}
 
 	public int deleteMovie(int movie_code) {
+		int existCount = movieMapper.checkExistSch(movie_code);
+		
+		if (existCount > 0) {
+			return -1;
+		}
+		
 		return movieMapper.deleteMovie(movie_code);
 	}
